Remove /spaces/ prefix from spaces
It was getting annoying to have to remember /spaces/, this makes it so that your urls are easier to remember.

We wanted to avoid a catchall route because then if you have an asset 404'ing in your html, you can enter infinite loops. What we're doing is matching any set of alphanumeric characters and slashes, but no punctuation marks.