Add color support to pen
This adds a setColor method to the pen, and includes the color in the payload that is sent to tape and the push server.

It also modifies canvas.drawSegment and canvas.drawLine to allow a color in the options hash.

Lastly, it adds a 'Black Pen', 'Red Pen', and 'Blue Pen' to the toolbelt.